Move defaults style into `style.js`
